backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
geoalchemy2==0.14.3

# Authentication
python-jose[cryptography]==3.3.0
httpx==0.26.0

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.1

# CORS
starlette==0.36.3

